refactor[w2.6]: notifications and feedback class have been moved to a new folder
