+++ HealthcareSystem/Backend/Controller/NotificationSurveyAndFeedback/FeedbackController.cs
@@ -5,6 +5,7 @@
  ***********************************************************************/
 
 using Backend.Repository;
+using Model.NotificationSurveyAndFeedback;
 using Model.Users;
 using Repository;
 using Service.NotificationSurveyAndFeedback;
