+++ HealthcareSystem/Backend/Model/MyDbContext.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Model.NotificationSurveyAndFeedback;
 using Model.Users;
 using System;
 using System.Collections.Generic;
