+++ HealthcareSystem/Backend/Model/Users/Patient.cs
@@ -5,6 +5,7 @@
  ***********************************************************************/
 
 using Model.Enums;
+using Model.NotificationSurveyAndFeedback;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
