+++ HealthcareSystem/Backend/Repository/FeedbackRepository/IFeedbackRepository.cs
@@ -1,4 +1,5 @@
-﻿using Model.Users;
+﻿using Model.NotificationSurveyAndFeedback;
+using Model.Users;
 using System;
 using System.Collections.Generic;
 using System.Linq;
