+++ HealthcareSystem/DoctorFrontend/GenerateReport.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using Controller.NotificationSurveyAndFeedback;
+using Model.NotificationSurveyAndFeedback;
 using Model.Users;
 using System;
 using System.Collections.Generic;
