+++ HealthcareSystem/DoctorFrontend/MessageUnconfirmedDrug.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using Controller.NotificationSurveyAndFeedback;
 using Controller.UsersAndWorkingTime;
+using Model.NotificationSurveyAndFeedback;
 using Model.Users;
 using System;
 using System.Collections.Generic;
