+++ HealthcareSystem/DoctorFrontend/PatientFileTab.xaml.cs
@@ -14,6 +14,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using System.Configuration;
+using Model.NotificationSurveyAndFeedback;
 
 namespace WpfApp1
 {
