+++ HealthcareSystem/SecretaryFrontend/View/ExaminationViewByDoctor.xaml.cs
@@ -23,6 +23,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 using System.Configuration;
+using Model.NotificationSurveyAndFeedback;
 
 namespace ProjekatZdravoKorporacija
 {
