+++ HealthcareSystem/SecretaryFrontend/View/SearchExaminations.xaml.cs
@@ -4,6 +4,7 @@ using Controller.RoomAndEquipment;
 using Controller.UsersAndWorkingTime;
 using Model.Doctor;
 using Model.Manager;
+using Model.NotificationSurveyAndFeedback;
 using Model.Users;
 using ProjekatZdravoKorporacija.ModelDTO;
 using Syncfusion.Windows.Shared;
